Organize and cleanup explorer
