--- apps/explorer/src/components/pagination/Pagination.module.css
@@ -1,111 +1,0 @@
-@tailwind components;
-
-@layer components {
-    .basecontainer {
-        @apply inline-block h-[2.3rem]
-        items-center rounded border-[1px] border-solid
-        border-gray-50 bg-inherit stroke-gray-75
-        font-sans text-sm text-gray-75;
-
-        box-shadow: 0 1px 2px rgba(16 24 40 / 5%);
-    }
-
-    .desktopfooter .basecontainer {
-        @apply mx-1 px-2 py-2;
-    }
-
-    .under6footer .basecontainer {
-        @apply mx-[1vw] px-2 py-2 sm:mx-[4px];
-    }
-
-    .btncontainer {
-        @apply basecontainer inline-block cursor-pointer hover:bg-sui-light hover:stroke-gray-100 hover:text-gray-100;
-
-        transition: all 0.4s ease;
-    }
-}
-
-button.pagenumber,
-button.nointeract,
-.rhs > select {
-    @apply basecontainer;
-}
-
-button.pagenumber {
-    @apply select-none border-sui-dark text-sui-dark;
-
-    background: linear-gradient(0deg, #f4fbff, #f4fbff), #fff;
-}
-
-button.gone {
-    @apply opacity-30;
-}
-
-/* RHS stats styling */
-.rhs > div {
-    @apply mx-auto mt-2 text-sm capitalize text-gray-75 sm:mr-[12px] sm:mt-0;
-}
-
-.rhs {
-    @apply flex;
-}
-
-.rhs > section {
-    @apply mx-auto;
-}
-
-.rhs > select {
-    @apply ml-[1vw] hidden cursor-pointer overflow-visible rounded-[6px] border-r-[12px] border-transparent sm:block;
-
-    outline: 1px solid #e9eaeb !important;
-}
-
-/* Less than 6 case */
-.under6footer {
-    @apply mt-2 text-left text-center sm:grid sm:grid-cols-12;
-}
-
-.under6footer > div:first-child {
-    @apply block self-center sm:col-span-8 sm:justify-self-start;
-}
-
-.under6footer > div:nth-child(2) {
-    @apply items-center justify-self-end sm:col-span-4;
-}
-
-/* Desktop */
-
-.desktopfooter {
-    @apply hidden text-center sm:mt-2 sm:grid sm:grid-cols-12 sm:text-left;
-}
-
-.desktopfooter > div:first-child {
-    @apply inline-block sm:col-span-8 sm:block sm:self-center;
-}
-
-.desktopfooter > div:nth-child(2) {
-    @apply sm:col-span-4 sm:items-center sm:justify-self-end;
-}
-
-/* Mobile */
-
-.mobilefooter {
-    @apply block w-[85vw] text-sm sm:hidden;
-}
-
-.mobiletoprow {
-    @apply my-2 grid grid-cols-6 sm:hidden;
-}
-
-.mobiletoprow > button:nth-child(1),
-.mobiletoprow > button:nth-child(3) {
-    @apply col-span-1;
-}
-
-.mobiletoprow > button:nth-child(2) {
-    @apply col-span-4 border-none shadow-none;
-}
-
-.mobilebottomrow {
-    @apply grid grid-cols-2 gap-2;
-}
