+++ apps/explorer/src/pages/epoch/EpochTimer.tsx
@@ -1,7 +1,7 @@
 // Copyright (c) Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-import { useEpochProgress } from '~/pages/epochs/utils';
+import { useEpochProgress } from '~/pages/epoch/utils';
 import { ProgressCircle } from '~/ui/ProgressCircle';
 import { Text } from '~/ui/Text';
