+++ apps/explorer/src/pages/epoch/index.tsx
@@ -7,8 +7,8 @@ import { useQuery } from '@tanstack/react-query';
 import { useMemo } from 'react';
 import { useParams } from 'react-router-dom';
 
-import { CheckpointsTable } from '../checkpoints/CheckpointsTable';
-import { validatorsTableData } from '../validators/Validators';
+import { CheckpointsTable } from '../checkpoint/CheckpointsTable';
+import { validatorsTableData } from '../validators';
 import { EpochProgress } from './stats/EpochProgress';
 import { EpochStats } from './stats/EpochStats';
 import { ValidatorStatus } from './stats/ValidatorStatus';
