+++ apps/explorer/src/pages/validators/index.tsx
@@ -211,7 +211,7 @@ export function validatorsTableData(
     };
 }
 
-function ValidatorPageResult() {
+export default function ValidatorPage() {
     const { data, isLoading, isSuccess, isError } = useGetSystemState();
 
     const numberOfValidators = data?.activeValidators.length || 0;
@@ -373,5 +373,3 @@ function ValidatorPageResult() {
         </div>
     );
 }
-
-export { ValidatorPageResult };
